Experiment Results: The original code generated a new DropDown from scratch, going through each model to build a DD for each machine. This made the code O(len(machines)*len(models)).
I decided, then, to create a 'core_dropdown', and have it
pull a copy that for each machine. This brought the complexity down to O(len(models) + len(machines)). Much
better.

But doing this had the side effect of making all the DD's work. Only the 1st and 4th DD worked before.

Hypothesis:
It failed to generate the buttons at one point. The
program lagged or something, causing it to miss the
'generation sequence'

Experiment:

(Play with it)